updated login test case to handle 'already logged in' popup and redirect
